+++ frontend/src/app/modules/page/parts/model/parts.model.ts
@@ -84,11 +84,3 @@ export interface PartResponse {
 }
 
 export type PartsResponse = PaginationResponse<PartResponse>;
-
-export interface PartsCountriesMapResponse {
-  [index: string]: number;
-}
-
-export interface PartsCountriesMap {
-  [index: string]: number;
-}