+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Removed
 - Removed usage of add-license-header script in FE hook
+- Removed usage of map and map component 
 
 ## [3.2.0] - 2023-04-17